% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBaseGroup.R
\name{addBaseGroup}
\alias{addBaseGroup}
\title{addBaseGroup}
\usage{
addBaseGroup(map_object, group, visible = FALSE)
}
\arguments{
\item{map_object}{leaflet map}

\item{group}{character}

\item{visible}{\code{TRUE} or \code{FALSE}}
}
\value{
modified leaflet map
}
\description{
addBaseGroup
}
