% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGlPointOverlay.R
\name{addGlPointOverlay}
\alias{addGlPointOverlay}
\title{addGlPointOverlay}
\usage{
addGlPointOverlay(
  map_object,
  data = getMapData(map_object),
  color = NULL,
  stroke = NULL,
  palette = "inferno",
  group = "Points",
  crs = WGS84_GPS,
  visible = TRUE,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{map_object}{leaflet map object}

\item{data}{sf object}

\item{color}{formula, like \code{~ pmin(1.0, value)}; can reference things in \code{data}}

\item{palette}{like "inferno"; passed to \code{\link[colourvalues:colour_values_rgb]{colourvalues::colour_values_rgb()}}}

\item{group}{group name}

\item{crs}{probably shouldn't change}

\item{visible}{show when map loads?}

\item{verbose}{logical}

\item{popup}{formula, like \code{~ GEOID10}; defaults to using \code{\link[leafpop:popup]{leafpop::popupTable()}}}
}
\value{
modified \code{map_object}
}
\description{
addGlPointOverlay
}
