% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGlPolygonOverlay.R
\name{addGlPolygonOverlay}
\alias{addGlPolygonOverlay}
\title{addGlPolygonOverlay}
\usage{
addGlPolygonOverlay(
  map_object,
  data,
  color = NULL,
  palette = "inferno",
  popup = NULL,
  group = "Polygons",
  crs = WGS84_GPS,
  visible = TRUE,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{map_object}{leaflet map object}

\item{data}{sf object}

\item{color}{formula, like \code{~ pmin(1.0, value)}; can reference things in \code{data}}

\item{palette}{like "inferno"; passed to \code{\link[colourvalues:colour_values_rgb]{colourvalues::colour_values_rgb()}}}

\item{popup}{formula, like \code{~ GEOID10}; defaults to using \code{\link[leafpop:popup]{leafpop::popupTable()}}}

\item{group}{group name}

\item{crs}{probably shouldn't change}

\item{visible}{show when map loads?}

\item{verbose}{logical}
}
\value{
modified \code{map_object}
}
\description{
addGlPolygonOverlay
}
