% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOverlay.R, R/addAwesomeMarkerOverlay.R,
%   R/addCircleMarkersOverlay.R, R/addCirclesOverlay.R, R/addMarkerOverlay.R,
%   R/addPolygonOverlay.R, R/addPolylineOverlay.R
\name{addOverlay}
\alias{addOverlay}
\alias{addAwesomeMarkerOverlay}
\alias{addCircleMarkersOverlay}
\alias{addCirclesOverlay}
\alias{addMarkerOverlay}
\alias{addPolygonOverlay}
\alias{addPolylineOverlay}
\title{addOverlay}
\usage{
addOverlay(
  fun,
  map_object,
  data,
  group,
  color = "black",
  label = NULL,
  layer_id = NULL,
  visible = TRUE,
  weight = 1,
  opacity = 1,
  ...,
  verbose = getOption("verbose", default = FALSE)
)

addAwesomeMarkerOverlay(...)

addCircleMarkersOverlay(...)

addCirclesOverlay(...)

addMarkerOverlay(...)

addPolygonOverlay(...)

addPolylineOverlay(map_object, ..., color = "black", weight = 2, opacity = 1)
}
\arguments{
\item{fun}{leaflet function like \code{\link[leaflet:map-layers]{leaflet::addCircleMarkers()}}}

\item{map_object}{leaflet map}

\item{data}{passed to \code{fun}}

\item{group}{passed to \code{fun}}

\item{layer_id}{passed to \code{fun}}

\item{visible}{\code{TRUE} or \code{FALSE} (initial state)}

\item{...}{passed to \code{fun}}

\item{verbose}{display messages}
}
\description{
addOverlay
}
\section{Functions}{
\itemize{
\item \code{addAwesomeMarkerOverlay}: add AwesomeMarkers layer

\item \code{addCircleMarkersOverlay}: add CircleMarkers layer

\item \code{addCirclesOverlay}: add Circles layer

\item \code{addMarkerOverlay}: add Markers layer

\item \code{addPolygonOverlay}: add Polygon layer

\item \code{addPolylineOverlay}: add Polyline layer
}}

