% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOverlayGroup.R
\name{addOverlayGroup}
\alias{addOverlayGroup}
\title{addOverlayGroup}
\usage{
addOverlayGroup(map_object, group, visible = FALSE)
}
\arguments{
\item{map_object}{leaflet map}

\item{group}{character}

\item{visible}{\code{TRUE} or \code{FALSE}}
}
\value{
modified leaflet map
}
\description{
addOverlayGroup
}
