% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOverlays.R
\name{addOverlays}
\alias{addOverlays}
\title{addOverlays}
\usage{
addOverlays(
  map_object,
  data,
  fun,
  ...,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{map_object}{leaflet map}

\item{data}{grouped data}

\item{fun}{function to add a single overlay, taking an argument \code{group}}

\item{...}{passed to \code{fun}}

\item{verbose}{logical}
}
\value{
leaflet map, with overlays added
}
\description{
Helper for adding multiple overlays to a leaflet map at the same time, using grouped data.
Group names will be passed to \code{fun}, which should be a function that actually adds an overlay, e.g. \code{\link[=addPolygonOverlay]{addPolygonOverlay()}}.
}
