% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTileset.R
\name{addTileset}
\alias{addTileset}
\title{addTileset}
\usage{
addTileset(
  map_object,
  tileset,
  group,
  pane = "tileset",
  zIndex = 0,
  attribution = NULL,
  options = pathOptions(),
  ...,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{verbose}{}
}
\description{
addTileset
}
