% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTitle.R
\name{addTitle}
\alias{addTitle}
\title{addTitle}
\usage{
addTitle(
  map_object,
  title,
  group = "Title",
  opacity = 1,
  size = "16px",
  align = "center",
  style = list(opacity = "{opacity}", `font-size` = "{size}", `text-align` = "{align}",
    `background-color` = "white", position = "fixed !important", bottom = "0\%", left =
    "0\%", width = "100\%", padding = "0", margin = "0", `padding-top` = "6px",
    `padding-bottom` = "6px", `font-weight` = "normal"),
  position = "topleft",
  ...
)
}
\arguments{
\item{map_object}{leaflet map}

\item{title}{character}

\item{size}{character}

\item{position}{character}
}
\value{
modified leaflet map
}
\description{
addTitle
}
