% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseGroups.R
\name{getBaseGroups}
\alias{getBaseGroups}
\title{getBaseGroups}
\usage{
getBaseGroups(map_object)
}
\arguments{
\item{map_object}{leaflet map}
}
\value{
character
}
\description{
getBaseGroups
}
\seealso{
\code{\link[=setBaseGroups]{setBaseGroups()}}
}
