% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverlayGroups.R
\name{getOverlayGroups}
\alias{getOverlayGroups}
\title{getOverlayGroups}
\usage{
getOverlayGroups(map_object)
}
\arguments{
\item{map_object}{leaflet map}
}
\value{
character
}
\description{
getOverlayGroups
}
\seealso{
\code{\link[=setOverlayGroups]{setOverlayGroups()}}
}
