% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resetLayersControl.R
\name{resetLayersControl}
\alias{resetLayersControl}
\title{resetLayersControl}
\usage{
resetLayersControl(
  map_object,
  tileset = NULL,
  options = layersControlOptions(collapsed = TRUE),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
resetLayersControl
}
