% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBaseGroups.R
\name{setBaseGroups}
\alias{setBaseGroups}
\title{setBaseGroups}
\usage{
setBaseGroups(map_object, value)
}
\arguments{
\item{map_object}{leaflet map}

\item{value}{character}
}
\value{
map_object, with attribute \code{base_groups} set to \code{value}
}
\description{
setBaseGroups
}
\seealso{
\code{\link[=getBaseGroups]{getBaseGroups()}}
}
