% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOverlayGroups.R
\name{setOverlayGroups}
\alias{setOverlayGroups}
\title{setOverlayGroups}
\usage{
setOverlayGroups(map_object, value)
}
\arguments{
\item{map_object}{leaflet map}

\item{value}{character}
}
\value{
modified leaflet map
}
\description{
setOverlayGroups
}
\seealso{
\itemize{
\item \code{\link[=addOverlayGroup]{addOverlayGroup()}}
\item \code{\link[=getOverlayGroups]{getOverlayGroups()}}
}
}
