% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setViewBayArea.R
\name{setViewBayArea}
\alias{setViewBayArea}
\alias{setViewSFBA}
\title{Recenter and zoom map to fit SF Bay Area}
\usage{
setViewBayArea(
  leaflet_map,
  lng = NULL,
  lat = NULL,
  zoom = NULL,
  ...,
  verbose = getOption("verbose")
)

setViewSFBA(
  leaflet_map,
  lng = NULL,
  lat = NULL,
  zoom = NULL,
  ...,
  verbose = getOption("verbose")
)
}
\arguments{
\item{leaflet_map}{created with \code{leaflet()} or \code{leaflet_SFBA()}}
}
\description{
Recenter and zoom map to fit SF Bay Area
}
\section{Functions}{
\itemize{
\item \code{setViewSFBA}: (alias)
}}

\note{
just an alias
}
\seealso{
\code{\link[=leaflet_SFBA]{leaflet_SFBA()}}
}
