% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setViewBounds.R
\docType{methods}
\name{setViewBounds}
\alias{setViewBounds}
\alias{setViewBounds,leaflet,list-method}
\alias{setViewBounds,leaflet,sfc-method}
\alias{setViewBounds,leaflet,sf-method}
\alias{setViewBounds,leaflet,Spatial-method}
\title{Set bounds of a leaflet map}
\usage{
setViewBounds(map, obj, ...)

\S4method{setViewBounds}{leaflet,list}(map, obj, ...)

\S4method{setViewBounds}{leaflet,sfc}(map, obj, ...)

\S4method{setViewBounds}{leaflet,sf}(map, obj, ...)

\S4method{setViewBounds}{leaflet,Spatial}(map, obj, ...)
}
\arguments{
\item{map}{leaflet map object}

\item{obj}{\code{sf} or \code{Spatial} object, or a named list or vector}
}
\description{
Based on the bounding box of an object you supply.
}
