% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_leaflet.R
\name{write_leaflet}
\alias{write_leaflet}
\title{Write a leaflet map to disk}
\usage{
write_leaflet(
  leaflet_map,
  file = NULL,
  ...,
  remove_url = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{leaflet_map}{(object) created with [leaflet::leaflet()], [leaflet_SFBA()], etc.}

\item{file}{(character) ending in `.html` (but see Details)}

\item{...}{further arguments to [mapview::mapshot()]}

\item{remove_url}{(logical) passed to [mapview::mapshot()] if `file` doesn't end with `.html`}

\item{verbose}{(logical)}
}
\description{
Write a leaflet map to disk
}
\details{
This function relies on [mapview::mapshot()].
If `file` ends with an extensions other than `.html` ---
for example, `.png` --- then `mapshot()` will try to handle it.
But you may need to pass other arguments, like `width` and `height`.
Also, `remove_url` is `TRUE` by default (but you can change that).
}
