% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_leaflet.R
\name{write_leaflet}
\alias{write_leaflet}
\title{Write a leaflet map to disk}
\usage{
write_leaflet(
  leaflet_map,
  file = NULL,
  ...,
  remove_url = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{leaflet_map}{(object) created with \code{\link[leaflet:leaflet]{leaflet::leaflet()}}, \code{\link[=leaflet_SFBA]{leaflet_SFBA()}}, etc.}

\item{file}{(character) ending in \code{.html} (but see Details)}

\item{...}{further arguments to \code{\link[mapview:mapshot]{mapview::mapshot()}}}

\item{remove_url}{(logical) passed to \code{\link[mapview:mapshot]{mapview::mapshot()}} if \code{file} doesn't end with \code{.html}}

\item{verbose}{(logical)}
}
\description{
Write a leaflet map to disk
}
\details{
This function relies on \code{\link[mapview:mapshot]{mapview::mapshot()}}.
If \code{file} ends with an extensions other than \code{.html} ---
for example, \code{.png} --- then \code{mapshot()} will try to handle it.
But you may need to pass other arguments, like \code{width} and \code{height}.
Also, \code{remove_url} is \code{TRUE} by default (but you can change that).
}
